refactor: update import paths and keep naming convention
